aac: fix debug printf (confSize)

+++ modules/aac/aac.c
@@ -285,7 +285,7 @@ static int module_init(void)
 
 	debug("aac: Encoder configuration: conf=%w, "
 	     "frameLength=%u, inputChannels=%u\n",
-	     enc_info.confBuf, enc_info.confSize,
+	     enc_info.confBuf, (size_t)enc_info.confSize,
 	     enc_info.frameLength, enc_info.inputChannels);
 
 	re_snprintf(prm.config, sizeof(prm.config), "%w",
